requirements.txt
ansible==9.5.1
molecule==24.2.1
molecule-plugins==23.5.3
python-vagrant==1.0.0
pytest-testinfra==10.1.0
